Move rbox_bom computation into a nested module
The stable version of OpenSCAD appears to evaluate functions earlier in
the rendering process compared to the development version of OpenSCAD.
This was causing rbox_bom() to fail as the box part variables weren't
set up at function evaluation time. Moving the logic into a nested
module with the box part setup wrapper module outside of the nested
module appears to solve this issue.